declarative_agent_sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "declarative-agent-sdk"
version = "0.9.0"
description = "Declarative YAML-based SDK for building production-grade Google ADK agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Google ADK Samples Contributors"}
]
keywords = [
    "google-adk",
    "agents",
    "declarative",
    "yaml",
    "llm",
    "langgraph",
    "multi-agent",
    "workflows",
    "genai"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies
dependencies = [
    "google-adk>=0.1.0",
    "google-genai>=0.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "langgraph>=0.1.0",
    "tiktoken>=0.5.0",  # For token management
]

# Optional dependency groups
[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# vLLM provider support
vllm = [
    "litellm>=1.0.0",
]

# Enhanced tool support
tools = [
    "langchain-community>=0.1.0",  # For Tavily and other tools
]

# All optional features
all = [
    "declarative-agent-sdk[dev,vllm,tools]",
]

[project.urls]
Homepage = "https://github.com/mngaonkar/google-adk-samples"
Documentation = "https://github.com/mngaonkar/google-adk-samples/tree/main/sdk#readme"
Repository = "https://github.com/mngaonkar/google-adk-samples"
"Bug Tracker" = "https://github.com/mngaonkar/google-adk-samples/issues"

# Package discovery configuration
# The package is called "declarative_agent_sdk" and source files are in the current directory
[tool.setuptools]
packages = ["declarative_agent_sdk"]

[tool.setuptools.package-dir]
declarative_agent_sdk = "."

# Black formatting configuration
[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

# Ruff linting configuration
[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

# MyPy type checking configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
